+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 4.8
 - Move Studies Certificate title and Signatures HTML to Template in Transcripts.php, rosariosis.sql, rosariosis_es.sql & rosariosis_fr.sql
 - Add TranscriptsIncludeForm() function in Transcripts.fnc.php & Transcripts.php
 - Add Transcripts header action hook in Transcripts.php
+- Add TranscriptsGenerate() & \_getTranscriptsStudents() functions in Transcripts.fnc.php & Transcripts.php
 
 Changes in 4.7.1
 ----------------